+++ src/components/SearchResultsList/SearchResultsList.test.js
@@ -4,7 +4,7 @@ import {
 import noop from 'lodash/noop';
 
 import Harness from '../../../test/jest/helpers/harness';
-import SearchResultsList from './search-results-list';
+import SearchResultsList from './SearchResultsList';
 import authorities from '../../../mocks/authorities.json';
 
 const renderSearchResultsList = (props = {}) => render(