+++ src/components/search-results-list/index.js
@@ -1,1 +1,0 @@
-export { default as SearchResultsList } from './search-results-list';