+++ src/constants/index.js
@@ -1,0 +1,3 @@
+export * from './searchableIndexesValues';
+export * from './rawSearchableIndexes';
+export * from './searchableIndexesMap';