+++ src/constants/searchableIndexesValues.js
@@ -1,0 +1,13 @@
+export const searchableIndexesValues = {
+  KEYWORD: 'keyword',
+  IDENTIFIER: 'identifier',
+  PERSONAL_NAME: 'personalName',
+  CORPORATE_CONFERENCE_NAME: 'corporateConferenceName',
+  GEOGRAPHIC_NAME: 'geographicName',
+  NAME_TITLE: 'nameTitle',
+  UNIFORM_TITLE: 'uniformTitle',
+  SUBJECT: 'subject',
+  CHILDREN_SUBJECT_HEADING: 'childrenSubjectHeading',
+  GENRE: 'genre',
+  AUTHORITY_UUID: 'authorityUUID',
+};