+++ src/hooks/useAuthorities/index.js
@@ -1,0 +1,1 @@
+export { default as useAuthorities } from './useAuthorities';