+++ src/routes/SearchRoute.js
@@ -1,0 +1,9 @@
+import { AuthoritiesSearch } from '../views';
+
+const SearchRoute = () => {
+  return (
+    <AuthoritiesSearch />
+  );
+};
+
+export default SearchRoute;